+++ commands/fun/dab.js
@@ -4,8 +4,6 @@ With one, the user can dab on another user EarthBound-style, and, with luck and 
 trusty random number generator, they can output up to 1000 damage!
 */
 
-const dabEmote = "<:marshDab:648374543005777950>";
-
 module.exports = {
   config: {
       name: "dab",
@@ -17,8 +15,9 @@ module.exports = {
   run: async (bot, message, args) => {
 
     // If there is no argument, just give the user a regular dab.
+    const dabEmote = "<:marshDab:648374543005777950>";
     if (!args || args.length < 1) {
-      if (message.guild.member(bot.user).hasPermission("ATTACH_FILES")) {
+      if (message.guild.me.hasPermission("ATTACH_FILES")) {
         return message.channel.send({files: ["./config/bot/media/marshDab.png"]});
       }
       else return message.channel.send(dabEmote);
@@ -33,38 +32,17 @@ module.exports = {
     // If the person decides to mention themself, replace the second mention with the word "themself".
     if (member.user == message.author) { member = "themself"; }
 
-    // initialize damage counter
-    let attackDamage = Math.floor(Math.random() * (Math.floor(1000) - Math.ceil(1) + 1) ) + Math.ceil(1);
+    const attackDamage = Math.floor(Math.random() * (Math.floor(1000) - Math.ceil(1) + 1) ) + Math.ceil(1);
+    const battleMessage = `• ${message.author} tried dabbing on ${member}! ${dabEmote}`;
+    let customText;
 
-    if (attackDamage == 1) {
-      return message.channel.send(`• ${message.author} dabbed on ${member}! ${dabEmote}\n` +
-       "•  . . .\n" + `• **${attackDamage} HP** of damage.`);
-    }
-    else if (attackDamage <= 50) {
-      return message.channel.send(`• ${message.author} dabbed on ${member}! ${dabEmote}\n` +
-        `• **${attackDamage} HP** of damage! \n` + "• (Wasn't very effective...)");
-    }
-    else if (attackDamage <= 100) {
-      return message.channel.send(`• ${message.author} dabbed on ${member}! ${dabEmote}\n` +
-        `• **${attackDamage} HP** of damage!`);
-    }
-    else if (attackDamage <= 450) {
-      return message.channel.send(`• ${message.author} dabbed on ${member}! ${dabEmote}\n` +
-        "• Oh, baby! \n" + `• **${attackDamage} HP** of damage!`);
-    }
-    else if (attackDamage <= 750) {
-      return message.channel.send(`• ${message.author} dabbed on ${member}! ${dabEmote}\n` +
-        "• *WOWZA!* \n" + `• **${attackDamage} HP** of damage!`);
-    }
-    else if (attackDamage <= 999) {
-      return message.channel.send(`• ${message.author} dabbed on ${member}! ${dabEmote}\n` +
-        "• ***SMAAAASH!!*** \n" + `• A whoppin' **${attackDamage} HP** of mortal damage!`);
-    }
-    else {
-      return message.channel.send(`• ${message.author} dabbed on ${member}! ${dabEmote}\n` +
-        "• ***OUCH!!*** \n" + `• **${attackDamage} HP** of mortal damage?!?! Talk about lucky!`);
-    }
+    if (attackDamage <= 25) { customText = `• It had no effect!`; }
+    else if (attackDamage <= 100) { customText = `• **${attackDamage} HP** of damage!`; }
+    else if (attackDamage <= 450) { customText = `• Oh, baby!\n• **${attackDamage} HP** of damage!`; }
+    else if (attackDamage <= 750) { customText = `• *WOWZA!*\n• **${attackDamage} HP** of damage!`; }
+    else { customText = `• ***SMAAAASH!!***\n• A whoppin' **${attackDamage} HP** of mortal damage!`; }
 
+    return message.channel.send(`${battleMessage}\n${customText}`);
   }
 }
 
@@ -73,8 +51,6 @@ function getUserFromMention(mention, guild) {
   const matches = mention.match(/^<@!?(\d+)>$/);
   // If supplied variable was not a mention, matches will be null instead of an array.
   if (!matches) return;
-  // However the first element in the matches array will be the entire mention, not just the ID,
-  // so use index 1.
-  const id = matches[1];
-  return guild.members.cache.get(id);
+  // However the first element in the matches array will be the entire mention, so use index 1.
+  return guild.members.cache.get(matches[1]);
 }